多种方式下线服务

+++ eureka-server/src/main/java/com/yu7/control/controller.java
@@ -1,19 +1,22 @@
 package com.yu7.control;
 
-import AllUtil.StringUtils.StringChange;
 import com.netflix.appinfo.InstanceInfo;
 import com.netflix.discovery.EurekaClient;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cloud.netflix.eureka.EurekaDiscoveryClient;
-import org.springframework.web.bind.annotation.*;
-import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
-import okhttp3.RequestBody;
 import okhttp3.Response;
+import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.netflix.eureka.EurekaDiscoveryClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
 import javax.annotation.Resource;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -31,15 +34,17 @@ public class controller {
     @Value("${eureka-server.name}")
     private String eurekaServer;
 
+
     private static final HashMap<Integer, String> sourceMap = new HashMap<>();
 
 
     /**
-     * 下线指定节点下的Eureka服务
+     * 下线指定节点下的Eureka服务，目前只是清除了Eureka的一级缓存
      * @return hello
      */
     @GetMapping("/service-down")
-    public String shutDown(@RequestParam Integer portParam,@RequestParam String vipAddress) {
+    public String shutDown(@RequestParam List<Integer> portParams,@RequestParam String vipAddress) {
+        List<Integer> successList = new ArrayList<>();
         //获取到服务名下的所有服务实例
         List<InstanceInfo> instances = eurekaClient.getInstancesByVipAddress(vipAddress, false);
         //map<端口-实例id>
@@ -53,26 +58,30 @@ public class controller {
         //创建请求体
         OkHttpClient client = new OkHttpClient();
 
-        //处理服务信息
-        String serviceInfo = sourceMap.get(portParam);
-        //创建请求去删除服务
-        Request request = new Request.Builder()
-                .url("http://"+eurekaServer+"/eureka/apps/" + serviceInfo)
-                .delete()
-                .build();
-        log.debug(request.url().toString());
-
-        try {
-            Response response = client.newCall(request).execute();
-            if (response.code() == 200) {
-                log.debug(serviceInfo+"服务下线成功");
-            } else {
-                log.debug(serviceInfo+"服务下线失败");
-            }
-        } catch (IOException e) {
-            log.error(e.toString());
+        if (ObjectUtils.isEmpty(portParams)){
+          return "端口为空"; //todo 完善自定义异常
         }
-
-        return "hello";
+        portParams.forEach(temp->{
+            //处理服务信息
+            String serviceInfo = sourceMap.get(temp);
+            //创建请求去删除服务
+            Request request = new Request.Builder()
+                    .url("http://"+eurekaServer+"/eureka/apps/" + serviceInfo)
+                    .delete()
+                    .build();
+            log.debug(request.url().toString());
+            try {
+                Response response = client.newCall(request).execute();
+                if (response.code() == 200) {
+                    log.debug(serviceInfo+"服务下线成功");
+                    successList.add(temp);
+                } else {
+                    log.debug(serviceInfo+"服务下线失败");
+                }
+            } catch (IOException e) {
+                log.error(e.toString());
+            }
+        });
+        return "goodbye service"+successList;
     }
 }
